feat: EntityMetadata 내에 class 타입을 갖도록 변경
